Configure the completion plugin
Now we are limiting the amount of options it gives us so we won't
have super large and anoying completion windows. Also we have set
a new max width and height to the documentation window to make it bigger
and easier to read and prompt text. We included borders as well.

Signed-off-by: Lincoln-Yuji <lincolnyuji@hotmail.com>