Add a macro for opening fzf inside floaterm WITHOUT the wrapper
The fzf floaterm wrapper is not customizable from config files and
I don't want to change the wrapper scripts. The default flags are
a lame choice so we use this macro to use our own alias defined
into my config files.

If fzf returns a path so we open the file, else we just close the
instance. notice that the instance is closed regardless of the value
returned is a path or empty.

Signed-off-by: Lincoln Yuji <lincolnyuji@hotmail.com>