Create a shell script to geerically desompress files
We are removing the extract function from the bashrc
to a separeted shell script. This is better since it does not
depend on if we are using zsh or bash as our interactive shell.

Also, having that as a shell script allows us to use it outside
the interactive shell tself, making it more versatile and
interacted to the rest of system.

Signed-off-by: Lincoln Yuji de Oliveira <lincolnyuji@hotmail.com>