nvim: Add buffer manager plugin
Managing and hopping between buffers in vim is painful.
This plugin makes our life easier and I believe it's more
efficient than using top-window tabs, since the name of
the file we are editting is already present in the status bar
and we don't want to see the name of other open buffers at
all times while coding.

Signed-off-by: Lincoln Yuji <lincolnyuji@hotmail.com>